Splits display in display + resource manager
